Math and Menu
Cleaned up
