--- Quiver/MATH_EXT.c
@@ -1,11 +1,0 @@
-#include "MATH_EXT.h"
-
-double DTR(double deg)
-{
-    return deg * M_PI / 180.0;
-}
-
-double RTD(double rad)
-{
-    return rad * 180.0 / M_PI;
-}
