+++ Quiver/MENU.c
@@ -1,5 +1,4 @@
-#include "MENU.h"
-#include <freeglut.h>
+#include "Menu.h"
 
 //------------------------------------------------------------------------
