+++ Quiver/MathExt.c
@@ -1,0 +1,11 @@
+#include "MathExt.h"
+
+double DTR(double deg)
+{
+    return deg * M_PI / 180.0;
+}
+
+double RTD(double rad)
+{
+    return rad * 180.0 / M_PI;
+}
